Derive Function highlighting from Identifier
The Function highlighting group seems to be used mostly for syntax that
has nothing to do with functions, e.g., the slashes in a file system
path when using the tree view browser or the "None" keyword in Python.
Therefore, it seems sane to simply use the default values for
hi-Function which is hi-Identifier.